Create global-monitoring.js
This script was created by Christian Thomas of SkyTruth for the purpose of monitoring environmental incidents around the world. This version of the script was finalized on 2017/07/06. This version is set up for no user interaction with the script; date range and position are both specified through the User Interface (UI). It provides users with the ability to search a variety of available satellite image sources:
    1. ESA - Sentinel 1
    2. ESA - Sentinel 2
    3. USGS - Landsat 7 * currently only RAW images, no TOA products.
    4. USGS - Landsat 8 * currently only RAW images, no TOA products.
    
Users are able to view all images in the above image collections for a specified date range, which intersect an AOI that is comprised of a 70,000 meters (70km) buffer of a specified center point by toggling the 'Image Collections'. Users may export a mosaic of the image collections clipped to the AOI, they may also view all scenes which intersect the AOI by toggling the 'Specific Scenes', and may export any of these scenes by clicking the corresponding 'Export Prep' button, allowing exports to be run from the 'Tasks' bar.